Change timeout to be of type unsigned long
We had a bug where timout becomes negative in the sendBuffer function because it's passed as a signed int. The default 60000 ms timeout overflows the range of a signed int on Arduino Uno (AVR). Int is 16 bits, thus 2^16 = 65536. Maximum values for a signed int is therefore 65536/2-1 = 32767. And we are trying to store 60000 into this. Therefore overflows to  -5535.

+++ src/Wisol.h
@@ -30,7 +30,7 @@
 
 const uint8_t WISOL_TX = 4;  //  Transmit port for For UnaBiz / Wisol Dev Kit
 const uint8_t WISOL_RX = 5;  //  Receive port for UnaBiz / Wisol Dev Kit
-const unsigned int WISOL_COMMAND_TIMEOUT = 60000;  //  Wait up to 60 seconds for response from SIGFOX module.  Includes downlink response.
+const unsigned long WISOL_COMMAND_TIMEOUT = 60000;  //  Wait up to 60 seconds for response from SIGFOX module.  Includes downlink response.
 
 class Wisol
 {
@@ -97,7 +97,7 @@ public:
 private:
   bool sendCommand(const String &cmd, uint8_t expectedMarkers,
                    String &result, uint8_t &actualMarkers);
-  bool sendBuffer(const String &buffer, int timeout, uint8_t expectedMarkers,
+  bool sendBuffer(const String &buffer, unsigned long timeout, uint8_t expectedMarkers,
                   String &dataOut, uint8_t &actualMarkers);
   bool setFrequency(int zone, String &result);
   uint8_t hexDigitToDecimal(char ch);
